Rename generic time icons (e.g. NeverGlyph) to TimeNeverGlyph, rename existing Time glyph to Clock-* glyph
