+++ src/icons/glyphs/clockAlwaysGlyph.ts
@@ -1,0 +1,5 @@
+import { ClockAlwaysGlyph, ClockAlwaysGlyphSecondary } from './svgs';
+import { useIcon, IconProps } from '../../shared-components/icon';
+
+export default (props: IconProps) =>
+  useIcon(ClockAlwaysGlyph, ClockAlwaysGlyphSecondary, props);
