+++ src/icons/glyphs/clockMostOfTheTimeGlyph.ts
@@ -1,0 +1,8 @@
+import {
+  ClockMostOfTheTimeGlyph,
+  ClockMostOfTheTimeGlyphSecondary,
+} from './svgs';
+import { useIcon, IconProps } from '../../shared-components/icon';
+
+export default (props: IconProps) =>
+  useIcon(ClockMostOfTheTimeGlyph, ClockMostOfTheTimeGlyphSecondary, props);
