+++ src/icons/glyphs/clockNeverGlyph.ts
@@ -1,0 +1,5 @@
+import { ClockNeverGlyph, ClockNeverGlyphSecondary } from './svgs';
+import { useIcon, IconProps } from '../../shared-components/icon';
+
+export default (props: IconProps) =>
+  useIcon(ClockNeverGlyph, ClockNeverGlyphSecondary, props);
