+++ src/icons/glyphs/clockSometimesGlyph.ts
@@ -1,0 +1,5 @@
+import { ClockSometimesGlyph, ClockSometimesGlyphSecondary } from './svgs';
+import { useIcon, IconProps } from '../../shared-components/icon';
+
+export default (props: IconProps) =>
+  useIcon(ClockSometimesGlyph, ClockSometimesGlyphSecondary, props);
