+++ src/icons/glyphs/index.ts
@@ -8,7 +8,6 @@ export { default as AcneOneGlyph } from './acneOneGlyph';
 export { default as AcneSevenGlyph } from './acneSevenGlyph';
 export { default as AcneThreeGlyph } from './acneThreeGlyph';
 export { default as AhaBhaTonerSolutionGlyph } from './ahaBhaTonerSolutionGlyph';
-export { default as AlwaysGlyph } from './alwaysGlyph';
 export { default as AntiAgingAndAcneGlyph } from './antiAgingAndAcneGlyph';
 export { default as AntiAgingGlyph } from './antiAgingGlyph';
 export { default as AzelaicAcidGlyph } from './azelaicAcidGlyph';
@@ -18,6 +17,10 @@ export { default as CheckBadgeGlyph } from './checkBadgeGlyph';
 export { default as ChestGlyph } from './chestGlyph';
 export { default as CleanserGlyph } from './cleanserGlyph';
 export { default as ClindamycinGlyph } from './clindamycinGlyph';
+export { default as ClockAlwaysGlyph } from './clockAlwaysGlyph';
+export { default as ClockMostOfTheTimeGlyph } from './clockMostOfTheTimeGlyph';
+export { default as ClockNeverGlyph } from './clockNeverGlyph';
+export { default as ClockSometimesGlyph } from './clockSometimesGlyph';
 export { default as CloggedPoresGlyph } from './cloggedPoresGlyph';
 export { default as CombinationGlyph } from './combinationGlyph';
 export { default as ConsultHistoryGlyph } from './consultHistoryGlyph';
@@ -53,9 +56,7 @@ export { default as MicellarWaterGlyph } from './micellarWaterGlyph';
 export { default as MixIngredientsGlyph } from './mixIngredientsGlyph';
 export { default as MoisturizerGlyph } from './moisturizerGlyph';
 export { default as MortorGlyph } from './mortorGlyph';
-export { default as MostOfTheTimeGlyph } from './mostOfTheTimeGlyph';
 export { default as NeckGlyph } from './neckGlyph';
-export { default as NeverGlyph } from './neverGlyph';
 export { default as NiacinamideGlyph } from './niacinamideGlyph';
 export { default as NoMakeupGlyph } from './noMakeupGlyph';
 export { default as NoneGlyph } from './noneGlyph';
@@ -71,7 +72,6 @@ export { default as RichMoisturizerGlyph } from './richMoisturizerGlyph';
 export { default as RocketshipGlyph } from './rocketshipGlyph';
 export { default as SelfieGlyph } from './selfieGlyph';
 export { default as SleepGlyph } from './sleepGlyph';
-export { default as SometimesGlyph } from './sometimesGlyph';
 export { default as StressGlyph } from './stressGlyph';
 export { default as SunGlyph } from './sunGlyph';
 export { default as SunblockGlyph } from './sunblockGlyph';
