--- src/icons/glyphs/mostOfTheTimeGlyph.ts
@@ -1,5 +1,0 @@
-import { MostOfTheTimeGlyph, MostOfTheTimeGlyphSecondary } from './svgs';
-import { useIcon, IconProps } from '../../shared-components/icon';
-
-export default (props: IconProps) =>
-  useIcon(MostOfTheTimeGlyph, MostOfTheTimeGlyphSecondary, props);
