--- src/icons/glyphs/svgs/MostOfTheTimeGlyph.tsx
@@ -1,21 +1,0 @@
-import * as React from 'react';
-
-function SvgMostOfTheTimeGlyph(props: React.SVGProps<SVGSVGElement>) {
-  return (
-    <svg
-      viewBox="0 0 48 48"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-      {...props}
-    >
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M16.73 17.036l-.382-1.26a3.198 3.198 0 001.905-1.823h-9v4.767h29.494v-4.767H35.32c-.405 1.437-1.453 2.608-2.855 3.083l-.381-1.26a3.198 3.198 0 001.904-1.823H19.584c-.404 1.437-1.453 2.608-2.855 3.083zm18.77-4.407H40V40H8V12.63h3.007c0-2.553 1.964-4.629 4.38-4.629 2.414 0 4.379 2.076 4.379 4.629h6.975c0-2.553 1.964-4.629 4.38-4.629 2.415 0 4.38 2.076 4.38 4.629zm-7.506 0c0-1.822 1.403-3.305 3.127-3.305 1.723 0 3.126 1.483 3.126 3.305h-6.253zM15.386 9.324c-1.723 0-3.127 1.483-3.127 3.305h6.254c0-1.822-1.403-3.305-3.127-3.305zM9.253 38.687h29.494V20.076H9.253v18.61zm7.767-6.26l-2.085-2.193.892-.943 1.193 1.271 3.558-3.76.891.932-4.45 4.692zm9.16 0l-2.085-2.193.892-.943 1.193 1.271 3.558-3.76.881.932-4.44 4.692z"
-        fill="currentColor"
-      />
-    </svg>
-  );
-}
-
-export default SvgMostOfTheTimeGlyph;
