+++ src/icons/glyphs/svgs/TimeAlwaysGlyph.tsx
@@ -11,7 +11,7 @@ function SvgTimeAlwaysGlyph(props: React.SVGProps<SVGSVGElement>) {
       <path
         fillRule="evenodd"
         clipRule="evenodd"
-        d="M29.064 8v4.2c5.563 1.957 9.566 7.254 9.556 13.49 0 7.886-6.414 14.31-14.31 14.31C16.424 40 10 33.586 10 25.69c0-6.147 3.904-11.385 9.349-13.41V8h9.715zm-1.236 1.235h-7.244v2.649a14.146 14.146 0 013.736-.514c1.206 0 2.382.168 3.508.455v-2.59zM11.235 25.69c0 7.204 5.86 13.075 13.075 13.075 7.214 0 13.075-5.87 13.075-13.075 0-7.205-5.87-13.075-13.075-13.075-7.204 0-13.075 5.87-13.075 13.075zm8.332-.01a4.745 4.745 0 019.487 0 4.745 4.745 0 01-9.487 0zm1.235.01a3.509 3.509 0 107.018-.002 3.509 3.509 0 00-7.018.002z"
+        d="M16.73 17.035l-.382-1.26a3.198 3.198 0 001.905-1.822h-9v4.767h29.494v-4.767h-3.429c-.404 1.437-1.453 2.608-2.854 3.082l-.381-1.26a3.198 3.198 0 001.904-1.822H19.584c-.404 1.437-1.453 2.608-2.855 3.082zm17.517-4.406H19.766C19.766 10.076 17.8 8 15.386 8s-4.38 2.076-4.38 4.629h1.253c0-1.822 1.403-3.305 3.127-3.305 1.724 0 3.127 1.483 3.127 3.305H8V40h32V12.63h-4.5C35.5 10.076 33.536 8 31.12 8c-2.415 0-4.379 2.076-4.379 4.629h1.253c0-1.822 1.403-3.305 3.127-3.305 1.723 0 3.126 1.483 3.126 3.305zM9.253 38.687h29.494V20.076H9.253v18.61zm2.866-8.453l2.085 2.192 4.44-4.692-.883-.932-3.557 3.76-1.193-1.271-.892.943zm9.16 0l2.085 2.192 4.44-4.692-.883-.932-3.557 3.76-1.203-1.271-.882.943zm10.282 2.192l-2.084-2.192.882-.943 1.202 1.271 3.558-3.76.882.932-4.44 4.692z"
         fill="currentColor"
       />
     </svg>
