+++ src/icons/glyphs/svgs/TimeAlwaysGlyphSecondary.tsx
@@ -9,11 +9,11 @@ function SvgTimeAlwaysGlyphSecondary(props: React.SVGProps<SVGSVGElement>) {
       {...props}
     >
       <path
-        d="M24 8C15.18 8 8 15.18 8 24s7.18 16 16 16 16-7.18 16-16S32.82 8 24 8zm0 30.5C16 38.5 9.5 32 9.5 24S16 9.5 24 9.5 38.5 16 38.5 24 32 38.5 24 38.5z"
+        d="M27.02 24.68c-.62 0-.65-.55-.65-.66 0-.15.04-.63.61-.63 3.1 0 6.71-2.18 6.71-8.34v-6.3a.74.74 0 00-.74-.75h-17.2c-.41 0-.75.34-.75.75v6.31c0 5.75 3.89 8.34 6.65 8.34.25 0 .43.06.55.19.1.1.15.25.15.41 0 .32-.21.7-.78.7-3.18 0-6.55 2.82-6.55 8.04v6.52c0 .41.34.75.75.75h17.2c.41 0 .75-.34.75-.75v-6.52c-.01-4.52-2.95-8.06-6.7-8.06zM16.51 38.5v-5.77c0-4.25 2.6-6.54 5.05-6.54 1.48 0 2.28-1.13 2.28-2.2 0-.55-.2-1.06-.57-1.45-.27-.29-.78-.64-1.62-.64-2.14 0-5.15-2.12-5.15-6.84V9.5h15.7v5.56c0 4.28-1.95 6.84-5.21 6.84-1.39 0-2.11 1.07-2.11 2.13 0 .55.19 1.07.54 1.47.4.45.97.7 1.61.7 2.91 0 5.19 2.87 5.19 6.54v5.76H16.51z"
         fill="currentColor"
       />
       <path
-        d="M24 14.33c.41 0 .75-.34.75-.75v-1.91c0-.41-.34-.75-.75-.75s-.75.34-.75.75v1.91c0 .41.34.75.75.75zM36.33 23.25h-1.91c-.41 0-.75.34-.75.75s.34.75.75.75h1.91a.749.749 0 100-1.5zM33.04 19.53c.13 0 .26-.03.37-.1l1.64-.95c.36-.21.48-.67.27-1.02a.751.751 0 00-1.02-.27l-1.64.95a.746.746 0 00.38 1.39zM28.83 15.63c.12.07.25.1.37.1.26 0 .51-.13.65-.38l.96-1.67a.746.746 0 00-1.29-.75l-.96 1.67c-.21.37-.09.83.27 1.03zM18.14 15.35c.14.24.39.38.65.38.13 0 .26-.03.37-.1.36-.21.48-.67.27-1.02l-.95-1.65a.742.742 0 00-1.02-.27c-.36.19-.48.64-.28 1l.96 1.66zM12.94 18.48l1.65.95c.12.07.25.1.37.1.26 0 .51-.13.65-.38.21-.36.08-.82-.27-1.02l-1.65-.95a.754.754 0 00-1.02.28c-.21.36-.09.82.27 1.02zM29.86 32.65a.742.742 0 00-1.02-.27c-.36.21-.48.67-.27 1.02l.96 1.66c.14.24.39.38.65.38.13 0 .26-.03.37-.1.36-.21.48-.67.27-1.02l-.96-1.67zM35.06 29.52l-1.65-.95a.746.746 0 00-.75 1.29l1.65.95c.12.07.25.1.37.1.26 0 .51-.13.65-.38.21-.35.09-.81-.27-1.01zM14.34 24c0-.41-.34-.75-.75-.75h-1.92c-.41 0-.75.34-.75.75s.34.75.75.75h1.92c.41 0 .75-.34.75-.75zM14.61 28.55l-1.67.96c-.36.21-.48.67-.27 1.02.14.24.39.38.65.38.13 0 .26-.03.37-.1l1.67-.96c.36-.21.48-.67.27-1.02a.75.75 0 00-1.02-.28zM19.17 32.37a.746.746 0 00-1.02.27l-.96 1.66a.746.746 0 00.64 1.12c.26 0 .51-.13.65-.38l.96-1.66c.21-.34.08-.8-.27-1.01zM24 33.67c-.41 0-.75.34-.75.75v1.91c0 .41.34.75.75.75s.75-.34.75-.75v-1.91c0-.41-.34-.75-.75-.75zM24 31.85a7.85 7.85 0 100-15.7 7.85 7.85 0 000 15.7z"
+        d="M25.75 27.08h-2.9c-2.87 0-5.2 2.33-5.2 5.2v4.25c0 .52.42.95.95.95H30c.52 0 .95-.42.95-.95v-4.25c0-2.87-2.33-5.2-5.2-5.2zM30.95 15.94v-4.25a.95.95 0 00-.95-.95H18.6a.95.95 0 00-.95.95v4.25c0 2.87 2.33 5.2 5.2 5.2h2.9c2.87.01 5.2-2.32 5.2-5.2z"
         fill="currentColor"
       />
     </svg>
