+++ src/icons/glyphs/svgs/TimeMostOfTheTimeGlyph.tsx
@@ -9,7 +9,9 @@ function SvgTimeMostOfTheTimeGlyph(props: React.SVGProps<SVGSVGElement>) {
       {...props}
     >
       <path
-        d="M29.064 12.2V8h-9.715v4.279c-.82.306-1.601.692-2.342 1.137-.366.217-.712.444-1.058.691a14.415 14.415 0 00-3.478 3.578c-.238.346-.455.712-.662 1.087-.9 1.62-1.493 3.43-1.71 5.347-.06.514-.099 1.037-.099 1.57v.09C10.05 33.596 16.404 39.95 24.221 40h.089c.534 0 1.058-.03 1.571-.089a14.151 14.151 0 005.357-1.71c.375-.207.731-.425 1.087-.662a14.588 14.588 0 003.577-3.479c.247-.336.475-.691.692-1.057a14.053 14.053 0 001.858-5.198c.08-.524.129-1.058.148-1.601.01-.178.03-.346.03-.524 0-6.226-4.003-11.523-9.566-13.48zm-8.48-2.965h7.244v2.59a14.226 14.226 0 00-3.508-.455c-.178 0-.356.01-.534.02-.543.02-1.067.069-1.59.158-.544.089-1.088.197-1.612.336V9.235zm-9.27 17.848l11.603 11.603c-6.098-.653-10.96-5.505-11.603-11.603zm13.411 11.662l-13.47-13.47c.06-2.016.573-3.914 1.453-5.604l17.61 17.611a13.02 13.02 0 01-5.593 1.463zm6.7-2.105L13.35 18.564a13.173 13.173 0 013.469-3.587l18.194 18.194a12.894 12.894 0 01-3.587 3.469zm4.27-4.536L17.886 14.305a13.09 13.09 0 015.416-1.65l14.043 14.043a13.093 13.093 0 01-1.65 5.406zm-10.663-19.46c6.63.367 11.957 5.683 12.323 12.324L25.032 12.645z"
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M16.73 17.036l-.382-1.26a3.198 3.198 0 001.905-1.823h-9v4.767h29.494v-4.767H35.32c-.405 1.437-1.453 2.608-2.855 3.083l-.381-1.26a3.198 3.198 0 001.904-1.823H19.584c-.404 1.437-1.453 2.608-2.855 3.083zm18.77-4.407H40V40H8V12.63h3.007c0-2.553 1.964-4.629 4.38-4.629 2.414 0 4.379 2.076 4.379 4.629h6.975c0-2.553 1.964-4.629 4.38-4.629 2.415 0 4.38 2.076 4.38 4.629zm-7.506 0c0-1.822 1.403-3.305 3.127-3.305 1.723 0 3.126 1.483 3.126 3.305h-6.253zM15.386 9.324c-1.723 0-3.127 1.483-3.127 3.305h6.254c0-1.822-1.403-3.305-3.127-3.305zM9.253 38.687h29.494V20.076H9.253v18.61zm7.767-6.26l-2.085-2.193.892-.943 1.193 1.271 3.558-3.76.891.932-4.45 4.692zm9.16 0l-2.085-2.193.892-.943 1.193 1.271 3.558-3.76.881.932-4.44 4.692z"
         fill="currentColor"
       />
     </svg>
