+++ src/icons/glyphs/svgs/TimeNeverGlyph.tsx
@@ -9,7 +9,7 @@ function SvgTimeNeverGlyph(props: React.SVGProps<SVGSVGElement>) {
       {...props}
     >
       <path
-        d="M29.064 12.2V8h-9.715v4.28C13.904 14.304 10 19.542 10 25.69 10 33.586 16.424 40 24.31 40s14.31-6.424 14.31-14.31c.01-6.236-3.992-11.533-9.556-13.49zm-8.48-2.965h7.244v2.59a14.225 14.225 0 00-3.508-.455c-1.295 0-2.54.188-3.736.514V9.235zm3.736 29.53c-7.214 0-13.075-5.87-13.075-13.075a13.07 13.07 0 012.965-8.292l18.718 18.105a12.98 12.98 0 01-8.608 3.262zm9.497-4.121L15.04 16.479a13.029 13.029 0 019.28-3.874c7.214 0 13.075 5.87 13.075 13.075 0 3.469-1.364 6.621-3.578 8.964z"
+        d="M35.5 12.629C35.5 10.076 33.536 8 31.12 8c-2.415 0-4.379 2.076-4.379 4.629h1.253c0-1.822 1.403-3.305 3.127-3.305 1.723 0 3.126 1.483 3.126 3.305H19.766C19.766 10.076 17.8 8 15.386 8s-4.39 2.076-4.39 4.629h1.253c0-1.822 1.403-3.305 3.127-3.305 1.724 0 3.127 1.483 3.127 3.305H8V40h32V12.629h-4.5zm3.247 26.058H9.253V20.076h29.494v18.61zm0-19.967H9.253v-4.767h8.99c-.351.847-1.023 1.525-1.905 1.822l.381 1.26c1.403-.476 2.445-1.641 2.846-3.082h14.402c-.351.847-1.022 1.525-1.904 1.822l.38 1.26c1.404-.476 2.446-1.641 2.847-3.082h3.437v4.767h.02z"
         fill="currentColor"
       />
     </svg>
