+++ src/icons/glyphs/svgs/TimeSometimesGlyph.tsx
@@ -11,7 +11,7 @@ function SvgTimeSometimesGlyph(props: React.SVGProps<SVGSVGElement>) {
       <path
         fillRule="evenodd"
         clipRule="evenodd"
-        d="M29.064 8v4.2c5.563 1.957 9.566 7.254 9.556 13.49 0 7.886-6.414 14.31-14.31 14.31C16.424 40 10 33.586 10 25.69c0-6.147 3.904-11.385 9.349-13.41V8h9.715zm-1.236 1.235h-7.244v2.649a14.146 14.146 0 013.736-.514c1.206 0 2.382.168 3.508.455v-2.59zM11.235 25.69c0 7.204 5.86 13.075 13.075 13.075 7.214 0 13.075-5.87 13.075-13.075 0-7.205-5.87-13.075-13.075-13.075-7.204 0-13.075 5.87-13.075 13.075zm7.877-3.532l.859-.888 9.858 9.537-.858.888-9.859-9.537z"
+        d="M31.11 8c2.416 0 4.38 2.076 4.38 4.63H40V40H8V12.63h2.997c0-2.553 1.964-4.629 4.38-4.629 2.414 0 4.379 2.076 4.379 4.63h6.975c0-2.554 1.964-4.63 4.38-4.63zm0 1.324c-1.723 0-3.126 1.483-3.126 3.305h6.253c0-1.822-1.403-3.305-3.126-3.305zm-15.734 0c-1.724 0-3.127 1.483-3.127 3.305h6.254c0-1.822-1.403-3.305-3.127-3.305zM9.253 38.687h29.494V20.076H9.253v18.61zm0-19.967h29.474v-4.767H35.29c-.401 1.44-1.444 2.606-2.846 3.082l-.381-1.26c.882-.297 1.553-.975 1.904-1.822H19.565c-.4 1.44-1.443 2.606-2.846 3.082l-.38-1.26c.881-.297 1.553-.975 1.903-1.822h-8.99v4.767zm9.44 8.082l.882.932-4.44 4.692-2.084-2.192.882-.943 1.203 1.271 3.557-3.76z"
         fill="currentColor"
       />
     </svg>
