+++ src/icons/glyphs/svgs/index.tsx
@@ -7,8 +7,6 @@ export { default as AcneOneGlyph } from './AcneOneGlyph'
 export { default as AcneSevenGlyph } from './AcneSevenGlyph'
 export { default as AcneThreeGlyph } from './AcneThreeGlyph'
 export { default as AhaBhaTonerSolutionGlyph } from './AhaBhaTonerSolutionGlyph'
-export { default as AlwaysGlyphSecondary } from './AlwaysGlyphSecondary'
-export { default as AlwaysGlyph } from './AlwaysGlyph'
 export { default as AntiAgingAndAcneGlyphSecondary } from './AntiAgingAndAcneGlyphSecondary'
 export { default as AntiAgingGlyphSecondary } from './AntiAgingGlyphSecondary'
 export { default as AzelaicAcidGlyphSecondary } from './AzelaicAcidGlyphSecondary'
@@ -22,6 +20,14 @@ export { default as ChestGlyph } from './ChestGlyph'
 export { default as CleanserGlyphSecondary } from './CleanserGlyphSecondary'
 export { default as CleanserGlyph } from './CleanserGlyph'
 export { default as ClindamycinGlyph } from './ClindamycinGlyph'
+export { default as ClockAlwaysGlyphSecondary } from './ClockAlwaysGlyphSecondary'
+export { default as ClockAlwaysGlyph } from './ClockAlwaysGlyph'
+export { default as ClockMostOfTheTimeGlyphSecondary } from './ClockMostOfTheTimeGlyphSecondary'
+export { default as ClockMostOfTheTimeGlyph } from './ClockMostOfTheTimeGlyph'
+export { default as ClockNeverGlyphSecondary } from './ClockNeverGlyphSecondary'
+export { default as ClockNeverGlyph } from './ClockNeverGlyph'
+export { default as ClockSometimesGlyphSecondary } from './ClockSometimesGlyphSecondary'
+export { default as ClockSometimesGlyph } from './ClockSometimesGlyph'
 export { default as CloggedPoresGlyphSecondary } from './CloggedPoresGlyphSecondary'
 export { default as CloggedPoresGlyph } from './CloggedPoresGlyph'
 export { default as CombinationGlyphSecondary } from './CombinationGlyphSecondary'
@@ -74,12 +80,8 @@ export { default as MoisturizerGlyphSecondary } from './MoisturizerGlyphSecondar
 export { default as MoisturizerGlyph } from './MoisturizerGlyph'
 export { default as MortorGlyphSecondary } from './MortorGlyphSecondary'
 export { default as MortorGlyph } from './MortorGlyph'
-export { default as MostOfTheTimeGlyphSecondary } from './MostOfTheTimeGlyphSecondary'
-export { default as MostOfTheTimeGlyph } from './MostOfTheTimeGlyph'
 export { default as NeckGlyphSecondary } from './NeckGlyphSecondary'
 export { default as NeckGlyph } from './NeckGlyph'
-export { default as NeverGlyphSecondary } from './NeverGlyphSecondary'
-export { default as NeverGlyph } from './NeverGlyph'
 export { default as NiacinamideGlyphSecondary } from './NiacinamideGlyphSecondary'
 export { default as NiacinamideGlyph } from './NiacinamideGlyph'
 export { default as NoMakeupGlyphSecondary } from './NoMakeupGlyphSecondary'
@@ -103,8 +105,6 @@ export { default as RocketshipGlyph } from './RocketshipGlyph'
 export { default as SelfieGlyphSecondary } from './SelfieGlyphSecondary'
 export { default as SelfieGlyph } from './SelfieGlyph'
 export { default as SleepGlyphSecondary } from './SleepGlyphSecondary'
-export { default as SometimesGlyphSecondary } from './SometimesGlyphSecondary'
-export { default as SometimesGlyph } from './SometimesGlyph'
 export { default as StressGlyphSecondary } from './StressGlyphSecondary'
 export { default as SunGlyphSecondary } from './SunGlyphSecondary'
 export { default as SunGlyph } from './SunGlyph'
